# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API LIVE TEST",
        "description": "Test API V1",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/v1/user/list": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "user_list",
                "operationId": "user_list",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 0
                    },
                    {
                        "name": "take",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 10
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/v1/user": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "user_store",
                "operationId": "userStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "mobile": {
                                        "type": "mobile",
                                        "format": "mobile",
                                        "example": "09123703808"
                                    },
                                    "email": {
                                        "type": "email",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "name"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": "name"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "family"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/v1/user/{user_id}": {
            "put": {
                "tags": [
                    "user"
                ],
                "summary": "user_update",
                "operationId": "userUpdate",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "mobile": {
                                        "type": "mobile",
                                        "format": "mobile",
                                        "example": "09123703808"
                                    },
                                    "email": {
                                        "type": "email",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": "name"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "family"
                                    },
                                    "change_password": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "user"
                ],
                "summary": "user_delete",
                "operationId": "userDelete",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/v1/user/profile": {
            "get": {
                "tags": [
                    "user_profile"
                ],
                "summary": "profile_show",
                "operationId": "profileShow",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/v1/user/profile/update": {
            "put": {
                "tags": [
                    "user_profile"
                ],
                "summary": "profile_update",
                "operationId": "profileUpdate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "name"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "family"
                                    },
                                    "lang_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "16272802467397.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/v1/user/profile/change_password": {
            "put": {
                "tags": [
                    "user_profile"
                ],
                "summary": "profile_change_password",
                "operationId": "profileChangePassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "change_password": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "password": {
                                        "type": "password",
                                        "example": "pass@1234"
                                    },
                                    "password_confirmation": {
                                        "type": "password",
                                        "example": "pass@1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Admin",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "Passport",
                "scheme": "bearer"
            },
            "bearerAuthSite": {
                "type": "http",
                "description": "Customer",
                "name": "AuthorizationSite",
                "in": "header",
                "bearerFormat": "Passport",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}